Revert "Merge pull request #4158 from cheesesashimi/zzlotnik/refactor-reconcile-code"
This reverts commit 1175d4030d525d75963c48acf86ef37445a843e4, reversing
changes made to 922db99998508b9fe8943737a845d6919b5b2630.
