+++ test/helpers/helpers.go
@@ -91,30 +91,27 @@ func NewMachineConfigExtended(
 	if annotations == nil {
 		annotations = map[string]string{}
 	}
-
-	ignConfig := &ign3types.Config{
-		Ignition: ign3types.Ignition{
-			Version: ign3types.MaxVersion.String(),
-		},
-		Storage: ign3types.Storage{
-			Files: files,
-		},
-		Systemd: ign3types.Systemd{
-			Units: units,
-		},
-		Passwd: ign3types.Passwd{},
-	}
-
-	if len(sshkeys) != 0 {
-		ignConfig.Passwd = ign3types.Passwd{
-			Users: []ign3types.PasswdUser{
-				{
-					Name:              "core",
-					SSHAuthorizedKeys: sshkeys,
+	rawIgnition := MarshalOrDie(
+		&ign3types.Config{
+			Ignition: ign3types.Ignition{
+				Version: ign3types.MaxVersion.String(),
+			},
+			Storage: ign3types.Storage{
+				Files: files,
+			},
+			Systemd: ign3types.Systemd{
+				Units: units,
+			},
+			Passwd: ign3types.Passwd{
+				Users: []ign3types.PasswdUser{
+					{
+						Name:              "core",
+						SSHAuthorizedKeys: sshkeys,
+					},
 				},
 			},
-		}
-	}
+		},
+	)
 
 	return &mcfgv1.MachineConfig{
 		TypeMeta: metav1.TypeMeta{
@@ -128,7 +125,7 @@ func NewMachineConfigExtended(
 		},
 		Spec: mcfgv1.MachineConfigSpec{
 			Config: runtime.RawExtension{
-				Raw: MarshalOrDie(ignConfig),
+				Raw: rawIgnition,
 			},
 			Extensions:      extensions,
 			FIPS:            fips,
